Add new keymap management on gh60
